==> Scripts/Tidy/3-health_dah.qmd <==
---
title : "Tidy - General health spending data - DAH"
editor: visual
date  : "March 3, 2023" 
---

### Purpose

-   Based upon `tidy-health_dah.R`, originally from November 17, 2021.

To tidy DAH general health expenditure data for China and the US.

### Setup

```{r}
#| warning: false
#| message: false
# ----------------------------------- #
# Clear working environment  
# ----------------------------------- #
rm(list = ls())
# ----------------------------------- #

# ----------------------------------- #
# Load required packages
# ----------------------------------- #
library(tidyverse)
library(sf)
library(countrycode)
# ----------------------------------- #

# ----------------------------------- #
# Load data
# ----------------------------------- #
panel <- read_rds("../../Data/tidy-data/0-panel.rds")
dah   <- read_csv("../../Data/source-data/Health-Aid_DAH/DAH_1990_2019_DATA.CSV")
# ----------------------------------- #
```

### Subset and summarize

Note on year range (2002-2017). There appears to be systematic missingness in Chinese aid prior to 2001 (131 countries per year) while 4 or fewer starting 2002. The former are therefore dropped while the latter are recoded as instances of no aid (aid = \$0).

In 2001 only Kosovo has a line-item allocation recorded in the DAH data. All other Chinese aid that year is classified as "Unallocated/Unspecified" motivating a 2002 start year.

This same problem reappears beginning 2018 - all aid values for both the US and China collapse to a single row reporting `Unallocated/Unspecified` recipient country.

**While there are `Unallocated/Unspecified` for the US as well as China, the vast majority of China's aid falls into this category within DAH.**

```{r}
#| warning: false

dah <- dah %>%
  filter(source  %in% c("China", "United_States"),
         year    %in% 2002:2017,
         # elim_ch: 1= Drop to avoid double counting transfers between channels
         elim_ch != 1) %>%
  mutate(dah_19     = as.numeric(dah_19) * 1e3,
         dah_hiv_19 = as.numeric(hiv_dah_19) * 1e3) %>%
  group_by(source, recipient_country,  year) %>%
  summarize(dah     = sum(dah_19, na.rm = TRUE),
            dah_hiv = sum(dah_hiv_19, na.rm = TRUE),
            .groups = "keep") %>%
  ungroup


# Subset panel here to identical year range for later join -- AND -- drop China / US
panel <- panel %>% 
  filter(year %in% 2002:2017,
         !ccode %in% c(2, 710)) %>% 
  st_drop_geometry() %>% 
  select(ccode, year)
```

\-\-\-\--

```{r}
# For Unallocated analysis - must do before unallocated dropped in ccode below
unal <- dah %>% 
  filter(recipient_country == "Unallocated/Unspecified") %>% 
  select(-dah_hiv) %>% 
  pivot_wider(., 
              id_cols     = year, 
              names_from  = source,
              values_from = dah)
```

\-\-\-\--

### IDs and pivot

Note - here also constructing estimate of `dah` less HIV support which is predominantly a USAID project not matched in China. Done in an attempt to resolve massive difference in the magnitude of values in these data.

```{r}
#| warning: false

# country codes
x <- tibble(
  recipient_country = sort(unique(dah$recipient_country)),
  ccode = countrycode(recipient_country, "country.name", "cown",
                      custom_match = c("Serbia" = 345))
)

# join country codes
dah <- left_join(dah, x, by = "recipient_country") %>% drop_na(ccode)

# relable source for tidy variable names post-pivot
dah <- dah %>% 
  mutate(source = case_when(source == "China"         ~ "cn",
                            source == "United_States" ~ "us"))

# construct dah measure that excludes hiv funding
dah <- dah %>% 
  mutate(dahnohiv = dah - dah_hiv) %>% 
  select(-dah_hiv)

# pivot wider
dah <- dah %>% 
  pivot_wider(., 
              id_cols = c(ccode, year), 
              names_from  = source,
              values_from = starts_with("dah"))

# Two instances of negative hiv aid for china recoded to 0
dah <- dah %>% 
  mutate(dahnohiv_cn = ifelse(dahnohiv_cn < 0, 0, dahnohiv_cn))

# Resolve missing observations in china aid
dah <- dah %>% 
  mutate(across(.cols = ends_with("cn"), 
                .fns  = ~replace_na(.x, 0)))

rm(x)
```

### Unallocated analysis

```{r}
aloc <- dah %>% 
  group_by(year) %>% 
  summarize(China_al         = sum(dah_cn, na.rm = T), 
            United_States_al = sum(dah_us, na.rm = T))

check <- left_join(unal, aloc, by = "year") %>% 
  mutate(per_cn = 100 * (China_al / (China + China_al)),
         per_us = 100 * (United_States_al / (United_States + United_States_al))) 

head(check %>% select(year, starts_with("per")))
tail(check %>% select(year, starts_with("per")))
summary(check %>% select(starts_with("per")))
```

```{r}
check <- check %>%
  mutate(us = United_States + United_States_al,
         cn = China + China_al) %>% 
  select(year, cn, us) %>% 
  mutate(across(.cols = !year, .fns = ~(.x/1e6))) %>% 
  mutate(across(.cols = !year, .fns = ~(.x - lag(.x, 1)), .names = "{.col}_fd")) %>% 
  pivot_longer(data = ., 
               cols = !year, 
               names_to  = "grp", 
               values_to = "val")

ggplot(data = check %>% filter(str_ends(grp, "_fd"))) +
  geom_line(aes(x = year, y = val, color = grp)) + 
  facet_wrap(~grp, ncol = 2, scales = "free_y")

ggplot(data = check %>% filter(!str_ends(grp, "_fd"))) +
  geom_line(aes(x = year, y = val, color = grp)) + 
  facet_wrap(~grp, ncol = 2, scales = "free_y")

rm(aloc, unal, check)
```

### Percent aid variables

Creating percent annual aid in another attempt to sidestep scale

```{r}
fx <- function(x){
  if(all(x == 0)){
    return(rep(0, length(x)))
  } else{
    return((x / sum(x)) * 100)  
    }
  }

dah <- dah %>% 
  arrange(ccode, year) %>% 
  group_by(year) %>% 
  mutate(across(.cols = starts_with("dah"), 
                .fns  = list("per" = fx))) %>% 
  ungroup %>% 
  select(ccode, year, sort(names(.)))

rm(fx)
```

### Merge to panel

Note - Dropping China observations from data - China has records giving health aid to itself(?) as well as receiving health aid from the United States.

```{r}
dah <- dah %>% filter(ccode != 710)

dah <- left_join(panel, dah, by = c("ccode", "year")) %>%
  mutate(across(.cols = starts_with("dah"),
                .fns  = ~replace_na(.x, 0)))

rm(panel)
```

### Save

```{r}
write_rds(x    = dah, 
          file = "../../Data/tidy-data/3-dah.rds")
```
